dashrequirement.txt
click==8.1.3
colorama==0.4.6
dash==2.8.0
dash-bootstrap-components==1.3.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-mantine-components==0.11.1
dash-table==5.0.0
Flask==2.2.2
gunicorn==20.1.0
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
Lasagne==0.1
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
numpy==1.21.6
pandas==1.3.5
plotly==5.13.0
protobuf==3.20.3
python-dateutil==2.8.2
pytz==2022.7.1
scikit-learn==1.0.2
scikit-neuralnetwork==0.7
scipy==1.7.3
six==1.16.0
tenacity==8.1.0
Theano==1.0.5
threadpoolctl==3.1.0
typing_extensions==4.4.0
Werkzeug==2.2.2
wincertstore==0.2
zipp==3.12.0
python-decouple==3.8
